pyproject.toml
[project]
name = "trackyr"
version = "0.1.0"
description = "Personal Windows desktop activity tracker"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "alembic>=1.13,<2.0",
    "pynput>=1.7,<2.0",
    "pystray>=0.19,<1.0",
    "Pillow>=10.0,<11.0",
    "psutil>=5.9,<7.0",
    "python-dotenv>=1.0,<2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
